MRMJ-1: Messenger & Skills fixes

+++ src/db/Cache.cpp
@@ -96,9 +96,9 @@ void CItemCache::OnFlush()
 		{
 			iLen += snprintf(szColumns + iLen, sizeof(szColumns) - iLen, ", socket0, socket1, socket2");
 			iValueLen += snprintf(szValues + iValueLen, sizeof(szValues) - iValueLen,
-					", %lu, %lu, %lu", p->alSockets[0], p->alSockets[1], p->alSockets[2]);
+					", %lu, %lu, %lu", static_cast<unsigned long>(p->alSockets[0]), static_cast<unsigned long>(p->alSockets[1]), static_cast<unsigned long>(p->alSockets[2]));
 			iUpdateLen += snprintf(szUpdate + iUpdateLen, sizeof(szUpdate) - iUpdateLen,
-					", socket0=%lu, socket1=%lu, socket2=%lu", p->alSockets[0], p->alSockets[1], p->alSockets[2]);
+					", socket0=%lu, socket1=%lu, socket2=%lu", static_cast<unsigned long>(p->alSockets[0]), static_cast<unsigned long>(p->alSockets[1]), static_cast<unsigned long>(p->alSockets[2]));
 		}
 
 		if (isAttr)
